Change parseInt to pasreFloat on api/products/edit

+++ src/app/api/products/edit.ts
@@ -94,7 +94,7 @@ export default async function edit(request: NextRequest) {
           id,
           categoryID,
           name,
-          price: parseInt(price),
+          price: parseFloat(price),
           intro,
           description,
           mainImageID: images[0].id,
